requirements.txt
psutil
psycopg2-binary
pandas
scikit-learn
joblib
python-dotenv